Do not return Z_BUF_ERROR in ceratain situations
Z_BUF_ERROR is to be returned if there is no input data available
for processing and therefore not ouput can be produced. Hardware
can keep input data in the scratch buffer. This causes the current
conditions to check for avail_in == 0 alone to be not sufficient,
such that we have cases were we do not extract all data. Disabling
the Z_BUF_ERROR return enables us to decode all data.