Avoid calling hardware if no new data did get send
This change avoids sending DDCBs when there is no data to be
processed.  This reduces the likelyhood of the situation which was
fixed in de1be2a37876cb89713aa4d9b5fb23e4d8942d5f Avoid overwriting
bytes in scratch buffer. Furthermore it improves the inflate
performance in particular for small data chunks.

Due to the software fallback if the data is below the inflate
threshold, this effect might be not noticable. Regardless, it is a
good idea to improve performance for small data chunks.