Fix negative avail_in problem
The circumvention code which checks for Z_STREAM_END conditions had
a problem calculating the right number of bytes to be taken from
next_in. This caused situations where negative values for avail_in
were returned, leading to problems in the calling code. The fact
that we take input bytes from the internal scratch buffer as well as
from the next_in buffer made the calculation tricky.