Change order for zlib.so dynamic loading
We are dynamically loading via dlopen(), dlsym() software zlib code
used to implement the fallback to software if there is no hardware or
if the requests are to small to perform well using hardware
decompression. We recommend using at least zlib 1.2.8. If a very old
distro is used it is possible to put a 1.2.8 or later in
/opt/genwqe/lib/libz.so which is prefered over the system libz.so.

This fix changes the order in which we try to find the libz.so.
ZLIB_PATH=... is the first which is tried, than the one in
/opt/genwqe/lib (which is normally not needed) and last the libz.so
from the distribution is taken.