Add more GenWQE specific tools
genwqe_vpdconf, genwqe_vpdupdate and csv2bin are tools used to read and
modify the card configuration data. Those tools are only to be used
by card manufacturers.

There is also a new genwqe-vpd target which will generate an RPM containing
the VPD altering tools.

genwqe_csum is a tool which calculates a zlib CRC32 using the hardware
accelerator. It is similar to genwqe_memcopy, but it does not transfer
the data back.