Add gzip/gunzip test-script
This script runs compression and decompresson requests and compares the
data with the original. It works in parallel such that the hardware
but also the software/driver is stressed a lot. The card can be selected
automatically or the load can be forced to one card only. Use -v to
get some messages out. Exit 0 means success anything else is failure.