Fix problem with installations not having a zlib-dev package
We have seen installations where there was no zlib-dev package containing
zlib.h to build the code. Our circumvention for this is to use a private
copy of zlib in /opt/genwqe/include|lib. Unfortunately the search path
to find those was only setup for the library and not for the tools.
This change should fix that. Unfortunately I was unable to remove my
zlib-dev on my development box due to important packages depending on it.
So someone needs to try out if this fix really works.