Fix problem with GenWQE specific optimization plus debug code to detect it
For GenWQE it matters how often the file-handle to the driver is opened
and closed. To reduce this for specific workloads, e.g. samtools Genomic
data processing, we avoided open/close as good as possible, sharing one
file handle for multiple threads. The CAPI version of our code is written
such that this won't work. But that is not serious, since we do not deal
with too many open/close calls, because each process has its own DDCB
queue in user-space.

This fix avoids the specific optimization in non-GenWQE cases.