Rework CAPI DDCB queue initialization and free-up
Fixed up the DDCB queue reference counter and locking during
AFU initializaton and free-up. The counter ctx->clients was
locked only when being increased, decreased. It was possible
to run into situations where the AFU was released accidently
two times leading to double free()s and follow up errors.
It was possible that ctx->clients become negative or inconsistent.

Statup and shutdown of the AFU are in symmetric functions now.
AFU, ddcb-queue and worker thread are started on the first usage
and freed when the usage count drops to 0. Buggy programs not
releasing the compression context will keep the worker thread and
AFU running until the process is entirely shutdown.