Added tracing in libddcb
The idea was to figure out how expensive the open, execution and close
operations are. We use this to compare the GenWQE with the CAPI version
of the hardware. One early result is that the CAPI hardware takes
significant time to close the AFU device. This impacts performance
a lot if not avoided by some means.