CAPI: Improve performance by never freeing AFU handle
Freeing the afu handle and stopping the associated thread takes
significant time. If an application should perform well, it must keep
the handle alive for the entire program livetime. We keep the thread,
the afu handle and all the associated memory. And now performance is
getting into an acceptable range. The ddcb execution itself was
already ok.