Do not cancel a thread which got never started
When using our gzip version with tar we got the folloing message:

tar: Child died with signal 11
tar: Error is not recoverable: exiting now

This was casued by a segfault in the destructor located in ddcb_capi.c.
Since we used a GenWQE card and not a CAPI card, we obviously tried to
stop something which was not allocated before.

Reported-by: Qing Wang <wangqing@cn.ibm.com>