Automatically download and build zlib-1.2.8
Using dlopen(), dlsym() to export software zlib functionality has turned
out to have deficiencies. One of them was recursively trying to dlopen()
itself if the LD_LIBRARY_PATH was setup but no /opt/genwqe/lib/zlib.so
was available. No error messages, no nothing. Another problem with the
approach is that if we found installations with old zlib version not providing
inflateGetDictionary(). Loading this might lead to undesired behaviour too.
So on the long-run, I like to statically include zlib, like we had it in
our early versions. This patch prepares for this.