Add printout of card utilization
To judge if the multicard feature is working nicely, I added
some code to dump more statistics e.g. which card executed
how many requests (DDCB).

In this patch, I also set the worker thread tid to 0, if the
init was failing. This avoids problem with pthread_cancel if
the thread is already gone because it failed to open the afu
e.g. if it is busy by another process in dedicated AFU mode.