Return inflate as soon as retc is not 0x102
We had a case where retc was not 0x102 and we continued to process
the data the ddcb returned. The code was trying to update the scratch
area and do some memmoves there with illegal size information.
retc != 0x102 means in this case to stop and not process further.
For GenWQE the code was correct, since the driver already checked retc.
For CAPI it did not. This fix is checking for retc and returns an error
if appropriate. This avoid heap corruption which we had seen for the CAPI
implementation before.