Fixup version information in the build system.
The trick to obtain the package version we have uses git describe
information, which is not always available, such as when we are building
the package from a tar.gz archive.  This patch fixes the build system to
look for the version information in a static file, version.mk, which can
be updated with the git version information when necessary.  If the file
is not available, make will either rely on git, or use the information
passed via the VERSION parameter in the command line interface.

This will be specially helpful to simplify packaging.