Implement signal handling free multi-card support
Imposing our libcard could freely use SIGIO for its recovery purpose
might not match what user-applications allow us to do. Therefore this
new version is polling every 4 sec to see if the allocated
file-descriptors still work nicely. If not, we close them and switch
to the alternate card, if there is any. Previously we used SIGIO to
trigger the validation process.