Allow owerwriting default location where to find libz.so when using dlopen mechanism
As discussed with Gabriel, we like to ensure that the libz.so of the
distribution we are running on is really being loaded via dlopen().
This should avoid trying to load ourselves accidently. I added
CONFIG_ZLIB_PATH which could be set now via make file option:

  make CONFIG_ZLIB_PATH="/usr/lib/libz.so"

I have not really tested that ;-). So we need to try if that works fine,
still.