Reintroduce zlib dependency
In addition I change the default libz.so search location, to what I think
is correct for the respective Linux distro. That way I like to prevent that
someone tries to load hw zlib by hw zlib, which leads to failure.
No need anymore to put a private zlib in /opt/genwqe/lib.