Add very useful testcase causing massive interrupt traffic and device closing
We are generating a lot of small requests, which are causing itself
interrupts on their completion. While doing that we kill all processes
suddenly in the middle of producing traffic. That is stressing the
interrupt handling code as well as the code which is responsible to
tidy up resources in the OS.