Build System: Remove version information malabarism.
My last patch series attempted to do some magic with the version
numbering, automatizing it in a way that both our git way and rpmbuild
would agree.  It has proven to have many flaws, and I don't think it's
worth fixing to solve a quite trivial problem.  Let's make it work
like it should be, requiring a small patch when updating version, but
simplifying a lot when doing packaging.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>