Build System: Use canonical name when uncompressing.
Use the canonical format <pkg>-<version> when uncompressing the file and
v<version>.tar.gz when fetching sources.  This one is caused by the way
github generates our releases.  This updates the make rpmbuild rules to
make it work with the new new but still comply with the spec file
"Source:" header.

The Git describe version hack needed an update because we don't want
a 'v' prefix before the version number in the specfile.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>