Build System: libcxl: Fixup bundling and installing from system path.
This commit refactors the bundling of libcxl so it won't get in the way
when we have an installation of it in the system path.  Now, we can
build using the bundle version by doing

$ make BUNDLE_LIBCXL=1

Or just use the system version, by simply calling

$ make

The bundle directory defaults to ext/libcxl, but if that needs to be
changed, one can say:

$ make BUNDLE_LIBCXL=1 CONFIG_LIBCXL_PATH=/path/to/libcxl

Be aware that default is not bundling, as it should be.  There is more
documentation on this on the config.mk file.  Read it for more
information on bundling libcxl.

For now, we choose the bundled version for our spec file, since libcxl
is not available in any distro and we can't redistribute it as a
separated package.

When we are bundling genwqe, let's build libcxl before even starting the
genwqe makefile.  This avoids some dependencies issues.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>