libcxl: Fixup bundling and installing from system path.
This commit refactors the bundling of libxl so it won't get in the way
when we have an installation of it in the system path.  Now, we can
build using the bundle version by doing

$ make BUNDLE_LIBCXL=1

Or just use the system version, by simply calling

$ make

The bundle directory defaults to ext/libcxl, but if that needs to be
changed, one can say:

$ make BUNDLE_LIBCXL=1 CONFIG_LIBCXL_PATH=/path/to/libcxl

Be aware that default is not bundling, as it should be.  There is more
documentation on this on the config.mk file.  Read it for more
information on bundling libcxl.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>