Build System: Remove duplicated installation of some tools.
For some reason zlib_mt_perf, genwqe_gzip, genwqe_gunzip, and friends
were getting installed twice, which messes up the packaging,
particularly zlib_mt_perf. In the case of genwqe_gzip and genwqe_gunzip,
it was getting installed on different directories, inside both /bin and
/bin/genwqe.  Since the later is forbidden in distros, we must stick
with only the /bin/genwqe_ version.  This was found when trying to
rename it zlib_mt_perf to comply with another request from downstream.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>