Fix falsely falling back to software mode
The code was querying the app_id and falsely returned a non-supported
value. That made the code drop to software mode in some
cases. Compression/decompression still worked, but was not as fast as
expected since software was being used.

We now return the requested value and basically delay reading out the
app_id. Also we use the full mask to ensure that no mixture between
different card version is used at the same time.