Remove cxl.h formerly used for systems not having uptodate kernel headers or using simulation on non powerpc systems
cxl.h is not licences under the Apache 2.0 license. Therefore having a
copy in this repository is not desired. Furthermore, it could even be
that there is a newer copy of the file already existing on kernel.org.
The problem with this file is that e.g. libcxl tried to do a wget to
download the file before building the code automatically. That is not
working in setups without internet access. Also it might just use whatever
version of this file. Removing it here, is the first step to get this right.